+++ src/receipts/receipt-form/receipt-form.component.ts
@@ -31,6 +31,7 @@ import {
 import { SnackbarService } from "../../services";
 import { QueueMode, ReceiptQueueService } from "../../services/receipt-queue.service";
 import { AuthState, FeatureConfigState, GroupState, UserState } from "../../store";
+import { downloadFile } from "../../utils/file";
 import { ItemListComponent } from "../item-list/item-list.component";
 import { UploadImageComponent } from "../upload-image/upload-image.component";
 
@@ -543,13 +544,7 @@ export class ReceiptFormComponent implements OnInit {
       .pipe(
         take(1),
         tap((blob) => {
-          const url = window.URL.createObjectURL(blob);
-          const link = document.createElement("a");
-          link.href = url;
-          link.setAttribute("download", currentImage.name); // or any other extension
-          document.body.appendChild(link);
-          link.click();
-          document.body.removeChild(link);
+          downloadFile(blob, currentImage.name);
         })
       )
       .subscribe();
