+++ src/receipts/receipts-table/receipts-table.component.html
@@ -14,6 +14,16 @@
       (afterClosed)="getFilteredReceipts()"
     >
     </app-quick-scan-button>
+    @if (dataSource.data.length > 0) {
+      <app-button
+        class="me-2"
+        icon="download"
+        matButtonType="iconButton"
+        tooltip="Export receipts"
+        color="accent"
+        (clicked)="exportReceipts()"
+      ></app-button>
+    }
     <app-button
       class="me-2 filter-button"
       icon="filter_alt"
@@ -34,16 +44,6 @@
       color="accent"
       (clicked)="resetFilterButtonClicked()"
     ></app-button>
-    @if (dataSource.data.length > 0) {
-      <app-button
-        class="me-2"
-        icon="export_notes"
-        matButtonType="iconButton"
-        tooltip="Export receipts"
-        color="accent"
-        (clicked)="exportReceipts()"
-      ></app-button>
-    }
     <ng-container *ngIf="canEdit && group?.groupSettings?.emailIntegrationEnabled">
       <app-button
         *appFeature="'aiPoweredReceipts'"
